Moved plotter to trading engine
