+++ src/core/adapters/yf_market_adapter.py
@@ -5,9 +5,7 @@ import yfinance as yf
 import tempfile
 from datetime import datetime, timedelta
 
-import plotly as py
-import plotly.graph_objects as go
-from plotly.subplots import make_subplots
+
 
 class YFMarketDataAdapter(MarketDataPort):
 
@@ -42,27 +40,3 @@ class YFMarketDataAdapter(MarketDataPort):
                 ]
             else:
                 raise NotImplemented("Asset Type not Implemented")
-    def generate_data_plot(self):
-
-        dates = [bar.timestamp for bar in self._list_data]
-        opens = [bar.open for bar in self._list_data]
-        highs = [bar.high for bar in self._list_data]
-        lows = [bar.low for bar in self._list_data]
-        closes = [bar.close for bar in self._list_data]
-
-        fig = make_subplots(specs=[[{"secondary_y": True}]])
-        fig.add_trace(go.Candlestick(
-            x=dates,
-            open=opens,
-            high=highs,
-            low=lows,
-            close=closes
-        ))
-        fig.update_layout(
-            title=f"Historical Data - {self.asset}", 
-            xaxis_title="Date", 
-            yaxis_title=f"{self.asset} Price ({self.asset.currency})",
-            xaxis_rangeslider_visible=False,
-            template="plotly_dark"
-            )
-        py.offline.plot(fig)
